requirements.txt
torch>=2.2.0
torchvision>=0.17.0
segmentation-models-pytorch>=0.5.0
numpy>=1.26.0
matplotlib>=3.10.0
opencv-python>=4.11.0
tqdm>=4.66.0
albumentations
Pillow>=10.0.0